requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
frozenlist==1.3.3
google-search-results==2.4.2
greenlet==2.0.2
idna==3.4
langchain==0.0.144
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.7
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.47
tenacity==8.2.2
termcolor==2.2.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
yaspin==2.3.0
